L4PROJ-17 cleaning up code base
