+++ main/app/src/App.tsx
@@ -2,15 +2,13 @@ import * as React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 import HomePage from './pages/home';
-import CaptionsPage from './pages/captions';
-import PeerPage from './pages/peer';
+// import CaptionsPage from './pages/captions';
 
 const App = () => (
   <Router>
     <Routes>
       <Route path="/" element={<HomePage />} />
-      <Route path="/captions" element={<CaptionsPage />} />
-      <Route path="/peer" element={<PeerPage />} />
+      {/* <Route path="/captions" element={<CaptionsPage />} /> */}
     </Routes>
   </Router>
 );
