+++ main/app/src/pages/home.tsx
@@ -2,34 +2,53 @@ import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Header from '../components/header';
 import "../styles/styles.css";
+import SpeechRecognition, { useSpeechRecognition } from 'react-speech-recognition'
+import "../styles/styles.css"
+import TextToSpeech from '../components/TextToSpeech';
+import AudioCapturePlayback from '../components/AudioPlayback';
 
 // let host = new Host();
 
 const HomePage = () => {
-  const [audio, setAudio] = useState<null | MediaStream>(null);
 
-  const navigate = useNavigate();
-  const handleClick = () => navigate('/captions');
+  // const navigate = useNavigate();
+  // const handleClick = () => navigate('/captions');
 
-  const getMicrophone = async () => {
-    const audio = await navigator.mediaDevices.getUserMedia({
-      audio: true,
-      video: false
-    });
-    setAudio(audio);
-  };
+  const startListening = () => {
+    SpeechRecognition.startListening({ continuous: true, language: 'en-IN' });
+  }
+  const {
+      transcript,
+      listening,
+      resetTranscript,
+      browserSupportsSpeechRecognition
+    } = useSpeechRecognition();
 
-  useEffect(() => {
-    console.log("audio", audio)
-  }, [audio])
+  if (!browserSupportsSpeechRecognition) {
+      return null
+  }
 
   return (
     <div>
       <Header title="Earbud Controller" showBackButton={false}/>
       <div className='pageBody'>
-        <button className="big-button" type="button" onClick={handleClick}>
-          Captions
-        </button>
+        <p className='small-text'>Listening: {listening ? 'on' : 'off'}</p>
+        
+        <div className='transcript-box'>
+            <div className='transcript-text'>
+                {transcript}
+            </div>
+        </div>
+
+        <div className="button-container">
+            <button className="small-button" onClick={startListening}>Start Listening</button>
+            <button className="small-button" onClick={SpeechRecognition.stopListening}>Stop Listening</button>
+            <button className="small-button" onClick={resetTranscript}>Reset</button>
+        </div>
+        
+        <TextToSpeech text={transcript} />
+
+        <AudioCapturePlayback />
       </div>
     </div>
   );
